Creation of store partiotions are now explicit

+++ packages/stash_sembast/lib/src/sembast/sembast_adapter.dart
@@ -3,7 +3,7 @@ import 'package:sembast/sembast_io.dart';
 import 'package:sembast/sembast_memory.dart';
 
 /// The [SembastAdapter] provides a bridge between the store and the
-/// Hive backend
+/// Sembast backend
 abstract class SembastAdapter {
   final int? version;
   final OnVersionChangedFunction? onVersionChanged;
@@ -13,7 +13,7 @@ abstract class SembastAdapter {
   /// The Sembast database object
   Database? _db;
 
-  /// List of stores per cache name
+  /// List of stores per name
   final Map<String, StoreRef<String, Map<String, dynamic>>> _stores = {};
 
   /// Builds a [SembastAdapter].
@@ -24,48 +24,73 @@ abstract class SembastAdapter {
   /// * [codec]: The codec which can be used to load/save a record, allowing for user encryption
   SembastAdapter({this.version, this.onVersionChanged, this.mode, this.codec});
 
+  /// Opens a Sembast database
   Future<Database> openDatabase();
 
-  Future<Database> _database() {
-    if (_db == null) {
-      return openDatabase().then((db) => _db = db);
+  /// Creates a store
+  ///
+  /// * [name]: The store name
+  Future<void> create(String name) {
+    Future<Database> database() {
+      if (_db == null) {
+        return openDatabase().then((db) => _db = db);
+      }
+
+      return Future.value(_db);
     }
 
-    return Future.value(_db);
+    if (!_stores.containsKey(name)) {
+      return database().then((db) {
+        _stores[name] = StoreRef<String, Map<String, dynamic>>(name);
+
+        return null;
+      });
+    }
+
+    return Future.value();
   }
 
-  StoreRef<String, Map<String, dynamic>> _store(String name) {
-    if (_stores.containsKey(name)) {
-      return _stores[name]!;
-    }
-
-    final store = _stores[name] = StoreRef<String, Map<String, dynamic>>(name);
-
-    return store;
+  /// Returns the [StoreRef]
+  ///
+  /// * [name]: The name of the store
+  ///
+  /// Returns the store [StoreRef]
+  StoreRef<String, Map<String, dynamic>>? _store(String name) {
+    return _stores[name];
   }
 
   /// Checks if a entry exists
   ///
-  /// * [name]: The cache name
-  /// * [key]: The cache key
+  /// * [name]: The store name
+  /// * [key]: The key
   ///
-  /// Returns true if the cache contains a entry with the provided key
+  /// Returns true if contains a entry with the provided key
   Future<bool> exists(String name, String key) {
-    return _database().then((db) {
-      return _store(name).record(key).exists(db);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.record(key).exists(db);
+    }
+
+    return Future.value(false);
   }
 
   /// Counts the number of entries using a optionally provided [Filter]
   ///
-  /// * [name]: The cache name
+  /// * [name]: The store name
   /// * [filter]: The [Filter]
   ///
   /// Returns the number of entries
   Future<int> count(String name, {Filter? filter}) {
-    return _database().then((db) {
-      return _store(name).count(db, filter: filter);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.count(db, filter: filter);
+    }
+
+    return Future.value(0);
   }
 
   /// Returns all the keys using a optionally provided [Finder]
@@ -75,21 +100,31 @@ abstract class SembastAdapter {
   ///
   /// Returns the number of entries
   Future<List<String>> keys(String name, {Finder? finder}) {
-    return _database().then((db) {
-      return _store(name).findKeys(db, finder: finder);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.findKeys(db, finder: finder);
+    }
+
+    return Future.value(const <String>[]);
   }
 
   /// Returns the json map associated with the provided key
   ///
-  /// * [name]: The cache name
-  /// * [key]: The cache key
+  /// * [name]: The store name
+  /// * [key]: The key
   ///
   /// Returns the key json map
   Future<Map<String, dynamic>?> getByKey(String name, String key) {
-    return _database().then((db) {
-      return _store(name).record(key).get(db);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.record(key).get(db);
+    }
+
+    return Future.value();
   }
 
   /// Returns the json maps of the provided keys
@@ -100,69 +135,93 @@ abstract class SembastAdapter {
   /// Returns the key json map
   Future<List<Map<String, dynamic>?>> getByKeys(
       String name, Iterable<String> keys) {
-    return _database().then((db) {
-      return _store(name).records(keys).get(db);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.records(keys).get(db);
+    }
+
+    return Future.value(const <Map<String, dynamic>?>[]);
   }
 
   /// Finds the list of all [RecordSnapshot]'s according with the provided
   /// [Finder]
   ///
-  /// * [name]: The cache name
+  /// * [name]: The store name
   /// * [finder]: The [Finder]
   ///
   /// Returns the list of records
   Future<List<RecordSnapshot<String, Map<String, dynamic>>>> find(String name,
       {Finder? finder}) {
-    return _database().then((db) {
-      return _store(name).find(db, finder: finder);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.find(db, finder: finder);
+    }
+
+    return Future.value(const <RecordSnapshot<String, Map<String, dynamic>>>[]);
   }
 
   /// Adds a value to the store
   ///
-  /// * [name]: The cache name
-  /// * [key]: The cache key
+  /// * [name]: The store name
+  /// * [key]: The key
   /// * [value]: The value to set
   /// * [merge]: If the value should be merged
   ///
   /// Returns the updated value
-  Future<Map<String, dynamic>> put(
-      String name, String key, Map<String, dynamic> value,
+  Future<void> put(String name, String key, Map<String, dynamic> value,
       {bool? merge}) {
-    return _database().then((db) {
-      return _store(name).record(key).put(db, value, merge: merge);
-    });
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store
+          .record(key)
+          .put(db, value, merge: merge)
+          .then((value) => null);
+    }
+
+    return Future.value();
   }
 
   /// Removes a entry by key
   ///
-  /// * [name]: The cache name
-  /// * [key]: The cache key
-  ///
-  /// Returns the dynamic record
-  Future<dynamic> remove(String name, String key) {
-    return _database().then((db) {
-      return _store(name).record(key).delete(db);
-    });
+  /// * [name]: The store name
+  /// * [key]: The key
+  Future<void> remove(String name, String key) {
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.record(key).delete(db);
+    }
+
+    return Future.value();
   }
 
   /// Clears the cache
   ///
   /// * [name]: The cache name
-  ///
-  /// Return the number of records updated
-  Future<int> clear(String name) {
-    return _database().then((db) {
-      return _store(name).delete(db);
-    });
+  Future<void> clear(String name) {
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.delete(db).then((value) => null);
+    }
+
+    return Future.value();
   }
 
   Future<void> _deleteStore(String name) {
-    if (_stores.containsKey(name)) {
-      return _database().then((db) {
-        return _store(name).delete(db);
-      }).then((_) => null);
+    final db = _db;
+    final store = _store(name);
+
+    if (db != null && store != null) {
+      return store.delete(db);
     }
 
     return Future.value();
@@ -182,30 +241,32 @@ abstract class SembastAdapter {
   /// named stashes in one storage, or stores(s) if multiple storages are used
   Future<void> deleteAll() {
     return Future.wait(_stores.keys.map((name) {
-      //return synchronized(() async {
-      return _database().then((db) {
+      final db = _db;
+
+      if (db != null) {
         return db
             .close()
             .then((_) => deleteDatabase(db.path))
             .then((_) => _stores.remove(name));
-      });
-      // });
+      }
+
+      return Future.value();
     }));
   }
 }
 
-class SembastPathAdapter extends SembastAdapter {
+class SembastLocalAdapter extends SembastAdapter {
   /// The location of the database file
   final String path;
 
-  /// Builds a [SembastPathAdapter].
+  /// Builds a [SembastLocalAdapter].
   ///
   /// * [path]: The location of the database file
   /// * [version]: The expected version
   /// * [onVersionChanged]:  If [version] not null and if the existing version is different, onVersionChanged is called
   /// * [mode]: The database mode
   /// * [codec]: The codec which can be used to load/save a record, allowing for user encryption
-  SembastPathAdapter(this.path,
+  SembastLocalAdapter(this.path,
       {int? version,
       OnVersionChangedFunction? onVersionChanged,
       DatabaseMode? mode,
